+++ src/App.tsx
@@ -11,6 +11,7 @@ import SettingsPage from './pages/SettingsPage';
 import OutputReportPage from './pages/OutputReportPage';
 import WeeklyReportPage from './pages/WeeklyReportPage';
 import ArchivePage from './pages/ArchivePage';
+import HelpPage from './pages/HelpPage';
 import { useTaskStore } from './store/useTaskStore';
 
 function App() {
@@ -40,6 +41,7 @@ function App() {
               <Route path="weekly" element={<WeeklyReportPage />} />
               <Route path="settings" element={<SettingsPage />} />
               <Route path="archive" element={<ArchivePage />} />
+              <Route path="help" element={<HelpPage />} />
             </Route>
           </Routes>
         </BrowserRouter>